First attempt at resolving issue #4
Updated the tests, altered the API and created a basic solution that
satisfied the new API.

Instead of extending SandwichStream, I simply created an instance of it
with the custom settings required. I'm tempted to leave it as is. I
don't think that I want to expose the full SandwichStream API to
MimePartStream users in case they break expectations.

In any case, 37 lines of code is most adequate.